# backend/docs/swagger.yaml
definitions:
...
  datatypes.JSONType-github_com_raids-lab_crater_dao_model_UserAttribute:
    type: object
...
  github_com_raids-lab_crater_dao_model.Role:
    enum:
    - 0
    - 1
    - 2
    - 3
    format: int32
    type: integer
    x-enum-varnames:
    - _
    - RoleGuest
    - RoleUser
    - RoleAdmin
...
  github_com_raids-lab_crater_dao_model.Status:
    enum:
    - 0
    - 1
    - 2
    - 3
    format: int32
    type: integer
    x-enum-comments:
      StatusActive: Active status
      StatusInactive: Inactive status
      StatusPending: Pending status, not yet activated
    x-enum-descriptions:
    - ""
    - Pending status, not yet activated
    - Active status
    - Inactive status
    x-enum-varnames:
    - _
    - StatusPending
    - StatusActive
    - StatusInactive
...
  github_com_raids-lab_crater_dao_model.UserBanRestrictions:
    properties:
      datasetDownload:
        type: boolean
      imageBuild:
        type: boolean
      jobSubmission:
        type: boolean
      modelDownload:
        type: boolean
      platformAccess:
        type: boolean
    type: object
...
  github_com_raids-lab_crater_internal_resputil.Page-internal_handler_UserResp:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_handler.UserResp'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  ? github_com_raids-lab_crater_internal_resputil.Response-github_com_raids-lab_crater_internal_resputil_Page-internal_handler_UserResp
  : properties:
      code:
        description: 依然保持 int (ErrorCode) 类型
        type: integer
      data:
        $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Page-internal_handler_UserResp'
      msg:
        type: string
    type: object
...
  internal_handler.UserResp:
    properties:
      attributes:
        allOf:
        - $ref: '#/definitions/datatypes.JSONType-github_com_raids-lab_crater_dao_model_UserAttribute'
        description: 用户额外属性
      banRestrictions:
        allOf:
        - $ref: '#/definitions/github_com_raids-lab_crater_dao_model.UserBanRestrictions'
        description: 当前生效的封禁限制，未封禁时为空
      banned:
        description: 当前是否处于封禁状态
        type: boolean
      bannedTimestamp:
        description: 用户封禁截止时间
        type: string
      extraBalance:
        description: 用户额外点数
        type: number
      id:
        description: 用户ID
        type: integer
      name:
        description: 用户名称
        type: string
      permanentBanned:
        description: 是否永久封禁
        type: boolean
      role:
        allOf:
        - $ref: '#/definitions/github_com_raids-lab_crater_dao_model.Role'
        description: 用户角色
      status:
        allOf:
        - $ref: '#/definitions/github_com_raids-lab_crater_dao_model.Status'
        description: 用户状态
    type: object
...
info:
  contact: {}
  description: This is the API server for Crater, a Multi-tenant AI Model Training
    Platform based on Kubernetes.
  title: Crater API
  version: 1.0.0
paths:
...
  /v1/admin/users/page:
    get:
      description: 按统一分页协议获取用户列表，搜索、筛选、排序和分页在数据库中执行
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size, 1-200
        in: query
        name: page_size
        type: integer
      - description: Search username or nickname
        in: query
        name: search
        type: string
      - description: Filter by role; repeatable
        in: query
        name: role
        type: integer
      - description: Filter by status; repeatable
        in: query
        name: status
        type: integer
      - description: Filter by current ban status; repeatable
        in: query
        name: banned
        type: boolean
      - description: Sort fields
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-github_com_raids-lab_crater_internal_resputil_Page-internal_handler_UserResp'
      security:
      - Bearer: []
      summary: 管理员获取用户分页
      tags:
      - User
...
securityDefinitions:
  Bearer:
    description: 访问 /login 并获取 TOKEN 后，填入 'Bearer ${TOKEN}' 以访问受保护的接口
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"